Remove dependency on dedis/crypto and other cleanups (#37)
If we ever want this benchmark again, we'll put it in dedis/kyber.

Added a new test to check the behavior of the String() method,
which "go vet" was complaining about. Changed format to include #
which does not make go vet less unhappy, but makes the string more
useful. Finally, go vet is giving a false positive on this situation.